Added gallows graphics to site
