+++ consts.js
@@ -1,0 +1,4 @@
+const GAME_STATE_FAILED = 'GAME_STATE_FAILED';
+const GAME_STATE_WON = 'GAME_STATE_WON';
+
+module.exports = { GAME_STATE_FAILED, GAME_STATE_WON };
