+++ package.json
@@ -31,7 +31,7 @@
     "webpack-cli": "^3.1.1"
   },
   "dependencies": {
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2"
+    "react": "^16.6.0",
+    "react-dom": "^16.6.0"
   }
 }
