+++ src/components/Alphabet/Letter/Letter.js
@@ -23,13 +23,13 @@ const checkLetter = (letter, id, onLetterClick) => {
     });
 }
 
-const Letter = ({letter, gameID, onLetterClick}) => {
+const Letter = React.memo(({letter, gameID, onLetterClick}) => {
     console.log("gameID",gameID)
     return(
         <div className={styles.letter} onClick={checkLetter.bind(null, letter, gameID, onLetterClick)}>
             {letter}
         </div>
     )
-}
+});
 
 export default Letter;
