+++ src/components/GameOver/GameOver.css
@@ -10,7 +10,7 @@
     border-radius: 20px;
 }
 
-.win {
+.gameOverFrame {
     font-size: 45px;
     font-weight: 600;
     padding: 40px;
@@ -18,6 +18,9 @@
     color: #2e8b57;
     text-align: center;
     white-space: pre-line;
+}
+.win {
+    color: #2e8b57;
 }
 .loose {
     color: #b22222;
